CHROMIUM: iwl7000: chromeOS: add $(srctree) in front of $(src) in the Makefiles
When compiling the generated kernel to a separate build directory, we
need to include the absolute path to the header files location.  But
we can't add $(srctree) in the backported tree (i.e. stack-dev),
because $(src) itself already contains the absolute path (and we even
remove any $(srctree) occurrences when creating the backport with
gentree.py).  To solve this, add $(srctree) (back) to the Makefiles in
the copy-code.sh script.

Change-Id: Idc0dad726be3f9a3f5386ec27ea56fc728b01fde
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 6ffd81c3a8b3f40ff2c11d0e8693fe8e39ea1d1d