CHROMIUM: Bluetooth: Add host controller identity as attribute
It is nice to have the controller identity available in sysfs so we can
correlate host controller with the connected peer device. While this is
available in debugfs, not all systems mount it nor give access to it to
anyone except root. This has been removed upstream as the same
information is available in debugfs. However, we would prefer it in
sysfs so we don't need to access debugfs from less privileged
applications.

BUG=b:146635705
TEST=cat /sys/class/bluetooth/hci0/identity and verified it is correct

Change-Id: I5f2bd960013d1afa476ae364ed79bab50d7ab500
Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1989879
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
(cherry picked from commit bda8ebbdfc9ba417b51492ceb97a9af5f8732903)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2068874
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>