CHROMIUM: iwl7000: don't throw error when trying to remove IGTK
The IGTK keys are only removed by mac80211 after it has already
removed the AP station.  This causes the driver to throw an error
because mac80211 is trying to remove the IGTK when the station doesn't
exist anymore.

The firmware is aware that the station has been removed and can deal
with it the next time we try to add an IGTK for a station, so we
shouldn't try to remove the key if the station ID is
IWL_MVM_INVALID_STA.  Do this by removing the check for mvm_sta before
calling iwl_mvm_send_sta_igtk() and check return from that function
gracefully if the station ID is invalid.

BUG=b:146861599
TEST=Connect and disconnect to an AP that uses IGTK and make sure an
    error message saying "offset 0 not used in fw key table" is not
    thrown.

Change-Id: I3d691c132920e84212e18069edb38b23494b0493
Cc: stable@vger.kernel.org # 4.12+
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 341511167c547c468f55a35744ad9dff6d3a9cc6
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1992382
Tested-by: Will Donnelly <wgd@google.com>
Reviewed-by: Will Donnelly <wgd@google.com>
Commit-Queue: Will Donnelly <wgd@google.com>