CHROMIUM: iwl7000: mvm: update the DTS measurement type
Till now, the driver asked the fw for a DTS measurement in automatic
mode. This triggered a flow in which the fw actively measured the
temperature.  This is not needed anymore, as the fw performs
measurements by itself, without the driver triggering them, and the
current cadence in which the fw performs such measurements is
sufficient.

In addition, in some time-sensitive scenarios, in which the driver asks
the fw for an active measurement twice in a short time (<100ms), the fw
asserts with code 0x20100801.

Change the DTS measurement to _WITHOUT_MEASURE instead, so the fw will
respond with the last measurement it has performed.

Change-Id: I53c9d76cc3ba946d7f93654ac7d6d0bd37b2ae31
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
(cherry picked from commit 3378fdf716eba1ba19898757a795bcee18d0b8b8)
iwl7000-tree: fc39fd0294c76a9debe9dcfc89d140af53753cb2