CHROMIUM: instrument pstore to debug crbug/999714
Look at the corrupted bytes.

Also, obtain the page frame number for this prz.
This should give us hints about if the overwrite
originates in the persistent RAM or elsewhere.

The hypothesis is that if the corruption originates
in the pstore module, then the PFN can be anywhere.
However, if it originates outside it, then the PFN
will probably be at the beginning of the persistent
memory range.

To be dropped once the bug is root-caused.

BUG=chromium:999714
TEST=Tested on fleex, including by removing the condition
and running the code unconditionally to verify that it is
harmless.
Change-Id: Id34f31c645858f29db1ef7ce8aa142b746a79bec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2001804
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>