CHROMIUM: iwl7000: chromeOS: replace NL80211_PREAMBLE_HE with a macro
The code in cfg80211 makes sure that no flag it doesn't understand is
included in the ftm preambles.  But iwl7000 now supports a new
preamble, namely NL80211_PREAMBLE_HE.  This was defined with the
correct value when it was backported, but it causes the bit to be set
and cfg80211 to reject it.

In order to prevent this, substitute the BIT(NL80211_PREAMBLE_HE) with
a macro that returns either the bit as it should be or 0 if the
cfg80211 version in use doesn't support it.

Change-Id: Ibce7df3d59ae6f7e126a9619a8c56534727dbfba
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit 3e1be24ae2cd10a2af0769981cdb3563c595bcd3)
iwl7000-tree: 0a58e9b5e37440251ceb0542e4decaf7949b46c9