CHROMIUM: drm: Make the drm_trace size configurable
Allow resizing of the ring buffer so we can fit in more than 2*PAGE_SIZE
buffers per-cpu.

Signed-off-by: Sean Paul <seanpaul@chromium.org>

Marked CHROMIUM since this patch isn't going to be posted upstream.
drm_trace won't be upstreamed as a part of drm, so it doesn't make sense
to post improvements to it.

BUG=chromium:1031787
TEST=Tested on fizz, changed the buffer size up and down

Change-Id: I8d0ed21261e4e8af94933a98942cba8f96d24568
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2003305
Reviewed-by: Rob Clark <robdclark@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>