CHROMIUM: iwl7000: mvm: add support for range request version 10
Range request version 10 keeps the same command size as version 9
but uses 2 reserved fields for the responder beacon interval and
station id (if exists).
For now, since the beacon interval of unassoc APs is unknown, use
a value of 100 TUs which is a common value for many APs.

While at it, remove the definition for CCMP_256 cipher, since this
is not supported.

Change-Id: I59ea7f773caed85a66c61401066ae169008442e6
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: bfc104f3eb3a97fc9637f6bb2c07d6c3498c9229