UPSTREAM: HID: wacom: Report 2nd-gen Intuos Pro S center button status over BT
The state of the center button was not reported to userspace for the
2nd-gen Intuos Pro S when used over Bluetooth due to the pad handling
code not being updated to support its reduced number of buttons. This
patch uses the actual number of buttons present on the tablet to
assemble a button state bitmap.

Link: https://github.com/linuxwacom/xf86-input-wacom/issues/112
Fixes: cd47de45b855 ("HID: wacom: Add 2nd gen Intuos Pro Small support")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Cc: stable@vger.kernel.org # v5.3+
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit dcce8ef8f70a8e38e6c47c1bae8b312376c04420)

BUG=chromium:1114790
TEST=compile tested, checked evtest events from CTL-672

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Change-Id: I702eb539fcd00047b3146454d2a965ad84373b2b