BACKPORT: FROMLIST: drm/i915: HDCP: retry link integrity check on failure
A single Ri mismatch doesn't automatically mean that the link integrity
is broken. Update and check of Ri and Ri' are done asynchronously. In
case an update happens just between the read of Ri' and the check against
Ri there will be a mismatch even if the link integrity is fine otherwise.

BUG=b:155304871
BUG=chromium:1080454
TEST=Build tested. Speculative based on failures in logs.

Signed-off-by: Oliver Barta <oliver.barta@aptiv.com>
Reviewed-by: Ramalingam C <ramalingam.c@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>

(am from https://patchwork.freedesktop.org/patch/364343/)

Downstream changes:
-intel_hdmi.c is in display/ upstream
-conflicts with hdcp2 code not in cros
-DRM_ERROR has been replaced by drm_err upstream

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I9d5b0acb26b5ea33e15142221243a37592e80f06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2210947
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>