UPSTREAM: Revert "rtw88: no need to set registers for SDIO"
This reverts commit 07d0f5534935e2daf63a4e1012af13d68e089fed.

For rtw88 driver, the SDIO is going to be supported, so there is
no need to remove the SDIO related power sequence settings. And
while the power sequence parser will pass in the mask of the HCI,
the SDIO part will not be used to set registers accordingly.

Moreover, the power sequence table is released as a whole package,
so the next time if we are going to update, the SDIO settings will
be overwritten. So, revert this now.

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20200520055350.23328-1-yhchuang@realtek.com
(cherry picked from commit dba5a189bf6169dc3e7462df17fb2aee5ca37e90)

BUG=none
TEST=none

Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
Change-Id: Ia1fb6baf78af241c1541c788b56d258ac29cc8dc
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2355779
Reviewed-by: Sean Paul <seanpaul@chromium.org>