CHROMIUM: iwl7000: chromeOS: make iwl7000 depend on x86_64 instead of all x86
ChromeOS had a patch that made iwl7000 depend on x86_64 or arm64.
Let's align with them and depend on x86_64 instead of x86, which means
both 32 and 64-bits.  We agreed not to support arm at all, so we don't
need to depend on arm64.

Change-Id: I8f3d8336c4d8fc6ee1993097f7b0a2f940a017b8
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 340d45d1264b119c486a1072b368a37ca192e11f