UPSTREAM: mfd: wm97xx-core: Platform data can be NULL
It is not mandatory that platform data is passed along the ac97
codec. Actually there are configuration without a battery connected to
the ADC of the codec.

This is for example the case for the PXA zylonite platform, so fix the
NULL dereferencing by adding a test.

Fixes: a5c6951c49fb ("mfd: wm97xx-core: core support for wm97xx Codec")
Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 46f107d0cb22b3f2b23d7fc7c50757f453a36508)

BUG=b:110025789
TEST=Test Audio use cases with full SOF 0.6 rebase release series applied.

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I624fc2f14dccc0f21afbaccb0e4f88d2fe6a728a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2259072
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>