CHROMIUM: mm: READ_ONCE(*pmd) before pmd_lock() in kstaled
The entry pointed to by pmd is unstable before it's locked. Accessing
the entry beforehand returns an invalid value when it races with
__split_huge_pmd(). Such race is rare but it crashes kernel because
pte_offset_map() tries to use the value as a pointer. Huge pud is not
supported for LRU pages but fix it too anyway.

BUG=b:123039911
TEST=Ran platform.MemoryPressure

Signed-off-by: Yu Zhao <yuzhao@google.com>
Change-Id: Ic7f3d0221651962d9988614994f23c6be3b5d94e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2274339
Reviewed-by: Salman Qazi <sqazi@google.com>
Commit-Queue: Yu Zhao <yuzhao@chromium.org>
Tested-by: Yu Zhao <yuzhao@chromium.org>