CHROMIUM: iwl7000: pcie: fix xtal latency for 9560 devices
We were using a very high latency for all 9560 devices so they all
would have time to stabilize.  But this causes the system to be
slighly slower, so we can use the best values for each device.

This requires a new trans cfg struct for devices with longer latency
and some adjustments to the other structs.

Change-Id: I3d3db14f6d1a8ecc547ca6afce8488816bd26081
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 88cc0b23aacbb4d53b359b4b0178153c19742cc4