CHROMIUM: iwl7000: mvm: remove memset of kek_kck command
iwl_mvm_wowlan_program_keys is now setting data directly in
the KEK_KCK command, and the memset is clearing this later,
causing the data to be incomplete.

Just remove the memory clearing, the structure is local and
cleared in the declaration.

Change-Id: Ifbb76a2814925f215a2c889c1792d46d5bd66ab8
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: 061abe53702a7204c3411789517b95f8b12545bb