UPSTREAM: usb: core: fix quirks_param_set() writing to a const pointer
The function quirks_param_set() takes as argument a const char* pointer
to the new value of the usbcore.quirks parameter. It then casts this
pointer to a non-const char* pointer and passes it to the strsep()
function, which overwrites the value.

Fix this by creating a copy of the value using kstrdup() and letting
that copy be written to by strsep().

Fixes: 027bd6cafd9a ("usb: core: Add "quirks" parameter for usbcore")
Signed-off-by: Kars Mulder <kerneldev@karsmulder.nl>

Link: https://lore.kernel.org/r/5ee2-5f048a00-21-618c5c00@230659773
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit b1b6bed3b5036509b449b5965285d5057ba42527)

BUG=b:117917134
TEST=rebuild kernel and apply usbcore.quirks into kernel cmdline,

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: Id236542180f1e17f2da8bf4c7f9fe7bf6bb0c53e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2339909
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@google.com>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>