CHROMIUM: iwl7000: mac80211: set sdata state once vif is up after restart
On FW recovery mac80211 is supposed to reconfigure the driver.
In the scanario when vif is of monitor type, after FW assert the vif
was getting added and channel context was getting created but this
channel context was not getting assigned to the vif.

This was fixed by setting the monitor sdata state bit to running once
the vif is up and running after the virtual monitor is added.

Change-Id: I494579f28018c2d0b9d4083a664cf872c28405ae
Signed-off-by: Borwankar, Antara <antara.borwankar@intel.com>
iwl7000-tree: fa16c79bd3eb7583d6a35a0002c8cc9977945c77