CHROMIUM: iwl7000: mvm: handle CCA-EXT delay firmware notification
If there are frequent CCA delays due to the extension channel
as detected by the firmware, and we're on 2.4 GHz, then handle
this by disconnecting (with a reconnect hint).

When we disconnect, we'll also update our capabilities to use
only 20 MHz on the next connection (if it's on 2.4 GHz) as to
avoid the use of the extension channel that has too much noise.

Change-Id: I709b7e6f73a7537c53f22d7418927691259de8a8
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: f668de980027d08034898090fb777f4dd323e997