CHROMIUM: iwl7000: skip offset when handling PNVM data section
IWL_UCODE_TLV_SEC_RT includes an offset to the SRAM address where the
data should be loaded, but in PNVM this is not needed, because we tell
the FW where to read the data from instead.

So we need to skip the offset that is present in the TLV and just copy
the actual data that comes after it.

Change-Id: I7eb4a5c6c3224b3a6b575d8cc566c23a120475c0
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: dc61bd33b9aaae931c7ff09e50e21fe7caeff5b3