CHROMIUM: iwl7000: fw api: make hdr a zero-size array again
In struct iwl_tx_cmd, there's no risk (as Arnd implied) that we
might access this as an array, as it's really not an array and
cannot be - there's only a single 802.11 header per frame. The
only reason for this member is for being able to access it a
bit more nicely.

On the other hand, this structure is used as a sub-struct in a
few places, and then some compilers (e.g. clang with certain
options) complain as you shouldn't have structs with variable-
length fields embedded in other structs.

Change-Id: I92179567d96938598806b560be59d787c2a8cc16
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: b19326645fa640f5876dad73bb31de68aa3c71df