CHROMIUM: iwlwifi: only support GeoSAR v0 tables
This cl reapplies the changes made in chromium:2513540 that were reverted
when merging in the Core57 changes (b:172642303)

Some existing GeoSAR tables incorrectly set the version field to 1
instead of 0 to mean that the contents of the table are invalid and
should be ignored.
However, recent versions of iwlwifi have started handling version 1 as
valid, resulting in incorrect GeoSAR tables being applied on those
devices.
Until we have audited and fixed those broken GeoSAR tables, keep
supporting only v0 tables.

BUG=b:172034758
TEST=load a v1 GeoSAR table that sets TX power to 0 on a device, observe
that the driver rejects the table and does not send the broken SAR table
to the firmware.

Change-Id: I95133baf20c46d465eebba6b1dddaf891a93cf4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2533381
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>