CHROMIUM: config: CONFIG_BOOTPARAM_HARDLOCKUP_PANIC=y
It's already set for non-x86 architectures on other kernels; x86 is the
oddball, instead relying on specifying `nmi_watchdog=panic` in
build_kernel_image.sh. It's generally best to avoid relying on kernel
command-line for this kind of configuration though, so specify it here
instead.

Generated by

  echo CONFIG_BOOTPARAM_HARDLOCKUP_PANIC=y >> \
      chromeos/config/x86_64/common.config
  chromeos/scripts/kernelconfig olddefconfig

atop a normalized config.

Note that kernel 4.14 doesn't let me enable this for arm/arm64;
fortunately, we have no ARM users on this kernel.

BUG=b:172222678
TEST=none

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: Idd46b11e21ade10bf7f399055a7d462076137335
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2530140
Reviewed-by: Stephen Boyd <swboyd@chromium.org>