CHROMIUM: iwl7000: copy iwl_he_capa for modifications
This data is not necessarily the same across devices as we may
modify it due to the number of antennas and for overrides (though
in practice overrides are likely to be identical), so modifying
the global data is wrong.

Make a copy of it in the NVM data and modify it there instead.

Change-Id: I50e063120e90ee02340653f14b6eb76a0cf9c769
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4a589f8dbdd34f6bdca1ddc2592305e09cf79c3c