CHROMIUM: iwl7000: pcie: correct lockdep annotation placement
When I migrated the code to use NAPI properly, I removed the same
lockdep annotations from the MSI-X interrupt handlers and instead
moved them into the NAPI processing.

This was wrong, because the sync_cmd_lockdep_map was intended to
protect us from accidentally taking the same lock while sending a
command to the firmware, and in the (threaded) interrupt handler,
so we'd deadlock since the interrupt handler couldn't tell anyone
that the command was finished.

This is all a pretty fringe scenario, but nonetheless the correct
placement of the lockdep annotations is in the interrupt handlers,
not in the NAPI poll function.

We cannot have a similar deadlock against the NAPI poll in the
first place since that is in soft-IRQ (or at least with soft-IRQs
disabled), and thus cannot take a lock that we can sleep under,
which is required for sending a sync command, so we don't need to
take any extra precautions for that case.

Change-Id: I942a08407282d9a0caab911d81b429bf1c3a22f6
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 0a2d0e1e90a0cf1e67578392e13938fe5e7f2eec