CHROMIUM: drm/i915: Fix corruption on octopus devices
We observed flickers on octopus devices and
including vblank_start in the vblank evasion stopped the
corruption (credit to David Stevens - stevensd@ for finding this).

As much as I try, I can't make sense of why this makes a
difference, and testing upstream kernel on octopus is not
possible so I can't test if this has been fixed upstream.
So this is a CHROMIUM patch for now.

Suggested-by: David Stevens <stevensd@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b/174528381,b/159775392
TEST=Tested on octopus, observed no flickers while watching TED app
     videos

Change-Id: I25f004197ca440f269eb3d29e72138079693a719
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2676205
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>