CHROMIUM: iwl7000: support loading the reduced power table from UEFI
This new feature allows OEMs to set a special reduced power table in a
UEFI variable, which we use to tell the firmware to change the TX
power tables.

Read the variable and store it in a dram block to pass it to the
firmware.  We do this as part of the PNVM loading flow.

Change-Id: I2e0bb142d2a9c412547cba89b62dd077b328fdc4
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: ea4e75a88f3467a064e949cdec1d863c82174a6d