FROMLIST: mwifiex: don't print SSID to logs
There are a few reasons not to dump SSIDs as-is in kernel logs:

1) they're not guaranteed to be any particular text encoding (UTF-8,
   ASCII, ...) in general
2) it's somewhat redundant; the BSSID should be enough to uniquely
   identify the AP/STA to which we're connecting
3) BSSIDs have an easily-recognized format, whereas SSIDs do not (they
   are free-form)
4) other common drivers (e.g., everything based on mac80211) get along
   just fine by only including BSSIDs when logging state transitions

Additional notes on reason #3: this is important for the
privacy-conscious, especially when providing tools that convey
kernel logs on behalf of a user -- e.g., when reporting bugs. So for
example, it's easy to automatically filter logs for MAC addresses, but
it's much harder to filter SSIDs out of unstructured text.

Signed-off-by: Brian Norris <briannorris@chromium.org>
(am from https://lore.kernel.org/patchwork/patch/1384122/)
(also found at https://lore.kernel.org/r/20210225024454.4106485-1-briannorris@chromium.org)

BUG=b:168844885
TEST=`dmesg`

Change-Id: Ibabd0eadae2a574811efafda9ec3dba8f46fa0e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2787222
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>