CHROMIUM: iwl7000: find reduced power table for the correct SKU
There was a misunderstanding in the original implementation of the
reduce power table reading format from UEFI.  The actual format
contains a list of tables with associated SKU_IDs.  We need to find
the correct SKU_ID and pass the corresponding table to the FW.
Instead, we were just passing the entire blob of data we got from the
UEFI variable.

Implement the TLV parsing part and pass only the payload that we
should pass.

Change-Id: I81bf1bdee64ae7e3f8b6c1a049b0adea6605d105
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 94e30fb3b51498479cda38dd3945c1aee1b757da