FIXUP: CHROMIUM: ASoC: rt5682: read property to get mclk control
0-day reports:

sound/soc/codecs/rt5682.c:2914:19: error:
	'struct rt5682_priv' has no member named 'mclk'; did you mean 'bclk'?

This is seen with CONFIG_COMMON_CLK=n.

Rebase note:

The patch introducing the problem was rejected upstream because it uses
a non-standard devicetree property name, not because the patch is
objectionable in general. The argument for keeping this a private patch
is therefore wrong and should be replaced with a more accurate comment
indicating that the patch was not accepted upstream because it uses an
unacceptable property name.

BUG=b:171755306
TEST=tested by OEM/AMD on Barla/Mordin

Change-Id: I1163f480f362cefa88cc756585f402f6306449d3
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2791604
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>