FROMGIT: Bluetooth: Return whether a connection is outbound
When an MGMT_EV_DEVICE_CONNECTED event is reported back to the user
space we will set the flags to tell if the established connection is
outbound or not. This is useful for the user space to log better metrics
and error messages.

BUG=b:173429036
TEST=Connect a classic and a LE mouse and check the flags values in
bluz.

Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Signed-off-by: Yu Liu <yudiliu@google.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 33091df75602beb36bb649e26fcbf07c9d934c82
 https://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git master)
Signed-off-by: Yu Liu <yudiliu@google.com>
Change-Id: I40a8d86f57ab014ee2f3c714f931f3dddbb6c4c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2840097
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Alain Michaud <alainm@chromium.org>