Revert "x86/build: Propagate $(CLANG_FLAGS) to $(REALMODE_FLAGS)"
This reverts commit d2b9d655f49294a32fdbc9499165989dfcacb636.

The patch causes clang internal assembler crashes with the clang
version used at Google.

chromeos-kernel-4_14-4.14.233-r1614:
	temp/video-bios-26d1d9.s: Assembler messages:
chromeos-kernel-4_14-4.14.233-r1614:
	temp/video-bios-26d1d9.s: Internal error (Segmentation fault).

This is necessary because the patch propagates CLANG specifics flags to
realmode flags, causing the respective code to be built with the internal
assembler.  Apparently that doesn't work with older kernel branches (the
same problem was seen in chromeos-4.19). Revert the patch to avoid using
the clang internal assembler in realmode for the time being. Maybe the
reverted patch can be reapplied after the clang assembler problem has been
fixed.

BUG=b:188943245
TEST=Build merge images with this patch reverted

Change-Id: Id61fbf9ba2fce73431ffa8e504264c50b21e1fd1
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@google.com>