CHROMIUM: ASoC: SOF: Intel: hda-dai: configure SSP in prepare stage
If a system suspend happens (e.g. with rtcwake in tests) while
playback/capture is on-going, the hw_params step is not invoked on
resume.

Other dais such as HDA and ALH/SoundWire use both the .hw_params and
the .prepare stage, but in the latter case use a state variable to
detect the difference between underflows and resume operations due to
restrictions on programming sequences.

For the SSP it's fine to only do the configuration in the .prepare
step.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>

BUG=b:187099195, b:192386991
TEST=Check audio playback on headphone path

Signed-off-by: Brent Lu <brent.lu@intel.com>
Change-Id: I23181c146aac8cf80ed121f202246ef21b4f3606
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2944492
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>