CHROMIUM: hid: Slice up 64-bit digitizer serial number
(This is intended to be be upstreamed, it is pending USB
HID standardization on a new Usage in this patch stack.)

Some existing HID devices report a 64-bit Digitizer.
Transducer Serial Number usage, which hid-input is
unable to process, and is not entirely HID spec
compliant.

As a step towards enabling processing of this field,
we specifically recognize such a 64-bit sized field
in hid-core, and slice the hid_field definition into
two separate 32-bit usages, the second being a
(proposed) new usage of Digitizer.Transducer Serial
Number Second 32 Bits.

This is not expected to affect compatibility, as
all existing APIs coming out of hid-core and
hid-input are unable to represent values larger
than 32 bits, and a 64-bit Transducer Serial Number
usage was already being truncated.

One noticeable difference to user-space is that
consumers of hidev, and debugfs hid information,
will see the transformed hid_field information
(with the sliced fields) instead of the unaltered
material from the original descriptor. Again, since
both paths truncate fields to 32 bits, no information
is actually being lost by this transformation.

BUG=b:184854727
TEST=Checked operation + power_supply functionality on entire patch stack; coachz, dedede

Signed-off-by: Kenneth Albanowski <kenalba@google.com>
Change-Id: Id5a239e8d8f8f8377508d60b42cd0812f6790d02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2945492
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2992998