CHROMIUM: iwl7000: dbg: treat non active regions as unsupported regions
If a region is not active, it means that it was not defined as a region
TLV in the FW image.  We should treat them as unsupported in that case.

This saves operational driver memory and run time when collecting debug
data by skipping unsupported regions.

Change-Id: I3ecb4e273bf714e426d82217e0590264cb763419
Signed-off-by: Rotem Saado <rotem.saado@intel.com>
iwl7000-tree: 49c96f8b018fc51353f77175e510e3789c9eb4d5