CHROMIUM: iwl7000: mac80211: do drv_reconfig_complete() before restarting all
When we reconfigure, the driver might do some things to complete
the reconfiguration. It's strange and could be broken in some
cases because we restart other works (e.g. remain-on-channel and
TX) before this happens, yet only start queues later.

Change this to do the reconfig complete when reconfiguration is
actually complete, not when we've already started doing other
things again.

For iwlwifi, this should fix a race where the reconfig can race
with TX, for ath10k and ath11k that also use this it won't make
a difference because they just start queues there, and mac80211
also stopped the queues and will restart them later as before.

Change-Id: Iefe494687f15fd85f77c1b989d1149c8efdfdc36
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 129679168ac38c77ec8020dfe7d70fa14aaa7309