FROMGIT: Bluetooth: Fix handling of experimental feature for quality reports
The existence of the experimental feature identifiy is the indication
that it is supported or not. No extra flag needed and the initial flag
should define if a feature is enabled or not. This is actually defined
in the management API definition.

Fixes: ae7d925b5c043 ("Bluetooth: Support the quality report events")
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
(cherry picked from commit 823f3bc4e2eca7170c2367c050c5361cb1cd7c25
git://git.kernel.org/pub/scm/linux/kernel/git/bluetooth/bluetooth-next.git master)
Signed-off-by: Joseph Hwang <josephsih@chromium.org>

BUG=b:203035116
TEST=Bluetooth quality report works.
For machines that do not support the quality report feature, the
messages log should show

  INFO bluetoothd: quality report feature not supported.

Change-Id: Id6f2afcde994f148c649a3019785292ea56b3e6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3225961
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>