CHROMIUM: iwl7000: mvm: Do not strip MIC when not needed
It is possible that the 'mic_crc_len' field also contains
the MIC which is expected to be handled by higher layer. Thus,
when removing the extra bytes do not remove the MIC if the HW
did not perform the decryption.

Change-Id: Iacda1d376e61e1aca581ab85d793c78f88ba5160
Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: abb705ba4f599992277769aabc9f6c481cf9917c