CHROMIUM: iwl7000: prefer WIDE_ID() over iwl_cmd_id()
The order of arguments for iwl_cmd_id() is confusing, and the
version is always 0 and thus a useless argument. Prefer the
WIDE_ID() macro (which needs to be a macro due to use in switch
cases etc.) over the iwl_cmd_id() function.

Obviously done with spatch:

  @@
  expression G, C;
  @@
  -iwl_cmd_id(C, G, 0)
  +WIDE_ID(G, C)

Change-Id: Ieb023cd773ea22e819d1ef1c37ae857ecc1a839d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 77657a3564761e41ec24524ae313e67c3fd8011c