drm: Implement `is_android_12` bool to fix HBM flickers
At the momemt, applying [1] and [2] fixes HBM flickers for FOD in android 12 but causes HBM flickers for FOD in android 11, not applying it therefore would be vice versa. Hence let's implement a bool, guard it with the snippets of code that is being executed to fix HBM flickers in android 12 and patch it from AnyKernel3.

Depends on [3].

[1]: https://github.com/RealJohnGalt/GaltsGulch-sm8150/commit/f0602ac1db024ece7fac446137c8d6320aaa7fdd
[2]: https://github.com/RealJohnGalt/GaltsGulch-sm8150/commit/1deeb7335b23cf45da44d2f855c21e9b2fefcb80
[3]: https://github.com/cyberknight777/AnyKernel3/commit/4e30cb19d63df4656a3a1f11dffe757fcc519910

Test: Compiled fine with GCC 12.0.0 and works fine but with an issue that the screen does not turn on even after using power button
Suggested-by: Adam W. Willis <return.of.octobot@gmail.com>
Suggested-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>