CHROMIUM: iwl7000: Fix MCC set country command
This command hasn't been working for a while, due to a change made in
the FW.  We need to use MCC_SOURCE_MCC_API to update the MCC and we need
to pass a key that tell the firmware that it's okay to change the MCC.
This is deprecated but still required.

This applies only for families above 9000.

BUG=b:208572707
TEST=chane MCC with internal Intel iw tool.

Change-Id: Ife27744799b2a7b5131081d9a8236dd30b90c939
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 6c9738e609040f85ca41b437d2cc66f2563f14c7
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3378064
Reviewed-by: Joshua Emele <jemele@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>