CHROMIUM: iwl7000: remove deprecated broadcast filtering feature
This feature has been deprecated and should not be used anymore.  With
newer firmwares, namely *-67.ucode and above, trying to use it causes an
assertion failure in the FW, similar to this:

[Tue Jan 11 20:05:24 2022] iwlwifi 0000:04:00.0: 0x00001062 | ADVANCED_SYSASSERT

In order to prevent this feature from being used, remove it entirely
and get rid of the Kconfig option that
enables it (IWLWIFI_BCAST_FILTERING).

Change-Id: Ia4f282813ca2ddd24c13427823519113f2bbebf2
Link: https://bugzilla.kernel.org/show_bug.cgi?id=215488
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 254485b268ffd0f8267ca7face79de510b55e6cf