kprofiles: Introduce KPROFILES_HELPER
Based on https://github.com/0ctobot/neutrino_kernel_oneplus_sm8150/commit/92d07930a99e06930ea43f3aaa1a1e09e634915b:

This is a simplified driver made to assist in injecting sepolicy for the sysfs nodes of kprofiles avoiding the need for root or a magisk module. This is achieved by temporarily disabling selinux for a split second for the userspace function to be executed.

Co-authored-by: celtare21 <celtare21@gmail.com>
Co-authored-by: Adam W. Willis <return.of.octobot@gmail.com>
[@cyberknight777: Adapt and refactor vbswap_helper as a dedicated kprofiles_helper]
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>